pyproject.toml
[project]
name = "legal-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastmcp>=2.11.0",
    "flask[async]>=3.1.1",
    "google-adk>=1.9.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.3",
]
